+++ integrated-scouting-apps/src/index.tsx
@@ -2,7 +2,7 @@
  * @author Nathan Kwok, add more...
  * @since 2023-10-29
  * @version 1.0.0
- * @description This is the entry point for the integrated-scouting-apps.
+ * @description welcome to tabata
  */
 import ReactDOM from 'react-dom/client';
 import HomePage from './routes/home';
@@ -12,6 +12,8 @@ import MatchScout from './routes/match';
 import DTF from './routes/dtf';
 import DTFTeams from './routes/dtfteams';
 import StrategicScout from './routes/strategic';
+import StrategicComment from './routes/strategicComment';
+import DriverSkill from './routes/driverSkill';
 import PitScout from './routes/pit';
 import Buh from './routes/buh';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
@@ -20,8 +22,7 @@ import TeamData from './routes/teamData';
 import Picklist from './routes/picklists';
 import Watchlist from './routes/watchlist';
 import WatchlistGet from './routes/watchlistdata';
-import WatchlistUpdate from './routes/watchlistupdate'
-
+import WatchlistUpdate from './routes/watchlistupdate';
 
 export default function App() {
   return (
@@ -32,15 +33,17 @@ export default function App() {
         <Route path="/scoutingapp" element={<ScoutingApp title="2637 Scouting App" />} />
         <Route path="/scoutingapp/match" element={<MatchScout title="2637 Match Scout" />} />
         <Route path="/scoutingapp/strategic" element={<StrategicScout title="2637 Strategic Scout" />} />
+        <Route path="/scoutingapp/strategic/comment" element={<StrategicComment title="2637 Strategic Scout" />} />
+        <Route path="/scoutingapp/strategic/driverskill" element={<DriverSkill title="2637 Driver Skill" />} />
         <Route path="/scoutingapp/pit" element={<PitScout title="2637 Pit Scout" />} />
         <Route path="/scoutingapp/lookup" element={<DataLookup title="2637 Data Lookup" />} />        
-        <Route path="/scoutingapp/lookup/teamData/:team_number" element ={<TeamData title="2637 Data"/>} />
-        <Route path="/scoutingapp/picklists" element={<Picklist title="2637 Picklists"/>} />
+        <Route path="/scoutingapp/lookup/teamData/:team_number" element ={<TeamData title="2637 Data Lookup" />} />
+        <Route path="/scoutingapp/picklists" element={<Picklist title="2637 Picklists" />} />
         <Route path="/dtf" element={<DTF title="2637 Drive Team Feeder" />} />
-        <Route path="/dtf/:team_number" element={<DTFTeams title="2637 Drive Team Feeders" />} />
+        <Route path="/dtf/:team_number" element={<DTFTeams title="2637 Drive Team Feeder" />} />
         <Route path="/watchlist" element = {<Watchlist title="2637 Watch List"/>}/>
-        <Route path="/watchlist/:team_number" element={<WatchlistGet title="2637 Watch List"/>} />
-        <Route path="/watchlist/update/:question_info" element={<WatchlistUpdate title="2637 Watch List"/>} />
+        <Route path="/watchlist/:team_number" element={<WatchlistGet title="2637 Watch List" />} />
+        <Route path="/watchlist/update/:question_info" element={<WatchlistUpdate title="2637 Watch List" />} />
         <Route path="/buh" element={<Buh/>} />
       </Routes>
     </BrowserRouter>
