+++ integrated-scouting-apps/src/public/stylesheets/watchlist.css
@@ -1,81 +1,16 @@
-/**
- * @author Robin Kang, add more...
- * @since 2024-2-28
- * @version 1.0.0
- * @description CSS for watchlist
- */
-
- .watchlistinput {
-    width: 100%;
-    margin-bottom:8%;
-    min-width: 100%;
-    height: 45% !important;
-    min-height: 45% !important;
-    border: 4px solid white !important;
-    border-radius: 25px;
-    background-color: transparent;
-    font-size: 500% !important;
-    font-family: Audiowide !important;
-    font-weight: bold;
-    color: white !important;
-    text-align: center !important;
-    background: transparent !important;
-
- }
- .watchlistBody {
-   font-size: 280%;
-   margin: 0;
-   margin-bottom:1%;
-   font-family: Audiowide !important;
-   background-color: transparent !important;
-   color: white;
-}
-.watchlistComment {
-   width: 90%;
-   margin-bottom: 8%;
-   min-width: 90%;
-   height: 20% !important;
-   min-height: 20% !important;
+.textbox_input {
+   width: 100%;
+   min-width: 100%;
+   height: 500px !important;
+   min-height: 500px !important;
    border: 4px solid white !important;
    border-radius: 25px;
-   background-color: transparent;
+   background-color: #433C3C;
    font-size: 500% !important;
-   font-family: Audiowide !important;
+   font-family: Figtree !important;
    font-weight: bold;
    color: white !important;
    text-align: center !important;
    background: transparent !important;
-}
-.watchlistbutton {
-   width: fit-content;
-   min-width: 50%;
-   height: 20% !important;
-   min-height: 20% !important;
-   border: 4px solid white !important;
-   border-radius: 100px;
-   background-color: transparent;
-   font-size: 350% !important;
-   font-family: Audiowide !important;
-   font-weight: bold;
-   color: white !important;
-   text-align: center !important;
-   margin-top: 5%;
-   padding-inline: 75px;
-}
-.undefinedbutton {
-   width: fit-content;
-   min-width: 50%;
-   height: 20% !important;
-   min-height: 20% !important;
-   border: 4px solid white !important;
-   border-radius: 100px;
-   background-color: transparent;
-   font-size: 350% !important;
-   font-family: Audiowide !important;
-   font-weight: bold;
-   color: white !important;
-   text-align: center !important;
-   margin-top: 5%;
-   padding-inline: 75px;
-   opacity: 0;
+   margin-bottom: 5%;
 }
